tests/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
coverage==7.3.2
exceptiongroup==1.1.3
execnet==2.0.2
fastapi==0.103.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
loguru==0.7.2
numpy==1.26.0
packaging==23.2
pandas==2.1.1
pluggy==1.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.2
pytest-cov==4.1.0
pytest-ordering==0.6
pytest-timeout==2.2.0
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
